[refactor] 리액트 컴포넌트 이름 교체
